removed monkeypatch.

+++ cogs/utils/site_save.py
@@ -1,5 +1,6 @@
 import os
 from time import time
+from typing import Optional
 import discord
 from .requests import fetch
 from functools import wraps
@@ -8,17 +9,20 @@ from functools import wraps
 def site_save(link):
     def deco(func):
         @wraps(func)
-        async def wrapper(self, *args, **kwargs):
+        async def wrapper(self, ctx, *args, **kwargs):
             try:
-                return await func(self, *args, **kwargs)
+                return await func(self, ctx, *args, **kwargs)
             except Exception as exception:
-                fname = f"data/runtime/crash_dump_{time()}.html"
-                with open(fname, "w") as crash_file:
-                    crash_file.write(await fetch(link))
-                self.bot.dev_logger.webhook_handler.webhook.send(
-                    file=discord.File(fname),
-                    username=self.bot.dev_logger.webhook_handler.username)
-                os.remove(fname)
+                fname: str = f"data/runtime/crash_dump_{time()}.html"
+                dump_channel: Optional[discord.TextChannel] = None
+                for channel in ctx.guild.text_channels:
+                    if channel.name == "marty_dev_log":
+                        dump_channel = channel
+                if dump_channel:
+                    with open(fname, "w") as crash_file:
+                        crash_file.write(await fetch(link))
+                    await dump_channel.send(file=discord.File(fname))
+                    os.remove(fname)
                 raise exception
 
         return wrapper
